refactor: Avoid deep nesting of imported files
