+++ packages/platform/src/containers/NotificationConfiguration/ConfigForm/components/List/types.ts
@@ -1,0 +1,15 @@
+import { ReactNode } from 'react';
+
+export type ListItemDetail = {
+  title: string;
+  description: string;
+  className?: string;
+};
+
+export type ListItem = {
+  title: string;
+  details?: ListItemDetail[];
+  description?: string;
+  titleClass?: string;
+  operations?: ReactNode;
+};
