+++ packages/platform/src/types.ts
@@ -7,13 +7,5 @@ export type NavItem = {
   authKey: string;
   tabs?: NavItem[];
   children?: NavItem[];
-  [key: string]: any;
-};
-
-export type ListItem = {
-  title: string;
-  details?: any;
-  description?: string;
-  titleClass?: string;
-  operations?: any;
+  [key: string]: unknown;
 };
